Made a lists class, optimized append method.
The append method is literally 100 times faster. Now the program keeps track of the last element in the list; so instead of iterating through the entire list until reaching the end to append an element, the program just appends the new element to the last element, and sets the last element to the appended element.